requirements.txt
faiss-gpu==1.7.2
h5py==2.10.0
lmdb==1.1.1
metric-learn==0.6.2
numpy==1.21.2
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
pandas==1.1.5
pytorch-ignite==0.1.2
PyYAML==5.4.1
scikit-image==0.16.2
scikit-learn==0.23.1
scipy
six==1.16.0
sklearn==0.0
timm==0.3.4
torch==1.8.0
torchvision==0.9.0
wrapt==1.12.1
yacs==0.1.7
